+++ src/MinEdLauncher/AuthorizedProduct.fs
@@ -20,3 +20,11 @@ let fixDirectoryPath productsDir platform directoryExists (product: AuthorizedPr
         | Frontier _ | Oculus _ | Dev ->
             if dirExists product.Sku then product.Sku else product.Directory
     { product with Directory = dir }
+    
+// Allows for overriding a product's filter. Useful for when FDev makes a copy/paste error
+// for a new product (i.e. when they released Odyssey with an "edh" filter instead of "edo") 
+let fixFilters overrides (product: AuthorizedProduct) =
+    overrides
+    |> Map.tryFind product.Sku
+    |> Option.map (fun filter -> { product with Filter = filter })
+    |> Option.defaultValue product